Split into 2 pages
Split our page into two pages: The Home page and Toolkit sub-page. The toolkit has the information we've worked on thus far, and the Home page is blank for now.
I removed the two-column layout from the Toolkit page, and all of the links/info that was in the second column. Don't worry, I saved a backup so if we need any of that text we still have it.
I changed the 3rd section (previously the Resource Guide section) from the Toolkit page and changed it to PDF Customization (the exact name I put was "Customize A Meet My Child Flyer"). We can always change the names later.
Changed the sub-headings in the text (the ones we said needed more emphasis) to h2 elements, and they do pop out better.
I FINALLY solved the button size problem on the toolkit page. Turns out, all I needed was a "style= width: 100%" tag.
I changed the names of the pages, but these names will have to be finalized later.
I fixed a small error with the main.js file loading before jQuery. I'm not sure how that got switched up, but I switched it back.
I've abandoned the PDF-Customization Git branch. It's most likely just going to go unused, but I'm keeping it up right now as a backup
Speaking of the PDF-Customization feature, I've kept it in a lightbox so far but we can decide if we want to move it to page content later.