Fixes from 10-12 Meeting
Changed the names of the pages to the names we decided on
I added some space between the purple buttons and the content underneath them. I did this using a style="padding-bottom:15px;" tag. It works, but I'm not sure if its quite the best solution. If we think of a better fix later we can go with that.
Removed the third section on the Toolkit page (and the corresponding button). Let me know what you think of the design changes, especially if the buttons are too big
Added buttons and sample text to the Pediatrics page
Did some research on how to implement inline citations in HTML. There is a citation tag, but it doesn't really do what I'm looking for, so I think the best option is a superscript link